Implement --single-user command line option
Passing this option will result in a fluent-bit instance to be spawned
on startup, with any events emitted to it being forwarded to vivo's
stderr (prefixed by "fluent-bit:").

When the option is passed, an additional "/console" endpoint is created,
which can be used to post data to the underlying fluent-bit instance.

Signed-off-by: Thiago Padilha <thiago@calyptia.com>